TestCafe lifecircle for multiwindow mode (#2181)
* initial

* fix form.submit page id

* cross-domain urls support

+++ src/client/utils/get-random-int-16-value.ts
@@ -1,8 +1,7 @@
 import nativeMethods from '../sandbox/native-methods';
 
 export default function (): number {
-    // @ts-ignore
-    const array = new nativeMethods.Uint32Array(1);
+    const array = new nativeMethods.Uint16Array(1);
 
     nativeMethods.cryptoGetRandomValues.call(nativeMethods.crypto, array);
